requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cycler==0.10.0
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==4.0
urllib3==1.26.7
yarl==1.7.0